optimized draw calls caused by cards
Normally, each Sprite3D requires a draw call. The exception: when the Sprite 3D does not contain any 3D transformations. Thus, we use a little trick in the handling of our cards: we make sure it's untransformed by swapping out front- and backside before turning the card around and ending the tween with a 'rotationY' of zero.